Move syntax module
