+++ omc4py/session/parser.py
@@ -7,7 +7,7 @@ __all__ = (
 
 import arpeggio  # type: ignore
 
-from . import syntax
+from omc4py.parser import syntax
 
 
 with syntax.omc_dialect_context:
